// src/models/user.ts
export interface User {
  id: number
  email: string
  name: Name
  status?: status
  phoneNumbers: string[]
}

export type status = "Happy" | "Sad"

export interface Name {
  first: string
  last?: string
}

export interface UserCreationRequest {
  email: string
  name: Name
  phoneNumbers: string[]
}